+++ kura/org.eclipse.kura.nm/src/main/java/org/eclipse/kura/nm/DeviceStateLock.java
@@ -16,6 +16,7 @@ import java.util.Objects;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
+import org.eclipse.kura.nm.enums.NMDeviceState;
 import org.freedesktop.dbus.connections.impl.DBusConnection;
 import org.freedesktop.dbus.exceptions.DBusException;
 import org.freedesktop.networkmanager.Device;
