+++ kura/org.eclipse.kura.nm/src/main/java/org/eclipse/kura/nm/enums/MMModemPowerState.java
@@ -10,7 +10,7 @@
  * Contributors:
  *  Eurotech
  *******************************************************************************/
-package org.eclipse.kura.nm;
+package org.eclipse.kura.nm.enums;
 
 import org.eclipse.kura.net.status.modem.ModemPowerState;
 import org.freedesktop.dbus.types.UInt32;
