+++ kura/org.eclipse.kura.nm/src/main/java/org/eclipse/kura/nm/status/NMStatusConverter.java
@@ -57,22 +57,22 @@ import org.eclipse.kura.net.status.wifi.WifiInterfaceStatus.WifiInterfaceStatusB
 import org.eclipse.kura.net.status.wifi.WifiMode;
 import org.eclipse.kura.net.status.wifi.WifiSecurity;
 import org.eclipse.kura.net.wifi.WifiChannel;
-import org.eclipse.kura.nm.MMBearerIpFamily;
-import org.eclipse.kura.nm.MMModem3gppRegistrationState;
-import org.eclipse.kura.nm.MMModemAccessTechnology;
-import org.eclipse.kura.nm.MMModemBand;
-import org.eclipse.kura.nm.MMModemCapability;
-import org.eclipse.kura.nm.MMModemLocationSource;
-import org.eclipse.kura.nm.MMModemMode;
-import org.eclipse.kura.nm.MMModemPortType;
-import org.eclipse.kura.nm.MMModemPowerState;
-import org.eclipse.kura.nm.MMModemState;
-import org.eclipse.kura.nm.MMSimEsimStatus;
-import org.eclipse.kura.nm.MMSimType;
-import org.eclipse.kura.nm.NM80211ApSecurityFlags;
-import org.eclipse.kura.nm.NM80211Mode;
-import org.eclipse.kura.nm.NMDeviceState;
-import org.eclipse.kura.nm.NMDeviceWifiCapabilities;
+import org.eclipse.kura.nm.enums.MMBearerIpFamily;
+import org.eclipse.kura.nm.enums.MMModem3gppRegistrationState;
+import org.eclipse.kura.nm.enums.MMModemAccessTechnology;
+import org.eclipse.kura.nm.enums.MMModemBand;
+import org.eclipse.kura.nm.enums.MMModemCapability;
+import org.eclipse.kura.nm.enums.MMModemLocationSource;
+import org.eclipse.kura.nm.enums.MMModemMode;
+import org.eclipse.kura.nm.enums.MMModemPortType;
+import org.eclipse.kura.nm.enums.MMModemPowerState;
+import org.eclipse.kura.nm.enums.MMModemState;
+import org.eclipse.kura.nm.enums.MMSimEsimStatus;
+import org.eclipse.kura.nm.enums.MMSimType;
+import org.eclipse.kura.nm.enums.NM80211ApSecurityFlags;
+import org.eclipse.kura.nm.enums.NM80211Mode;
+import org.eclipse.kura.nm.enums.NMDeviceState;
+import org.eclipse.kura.nm.enums.NMDeviceWifiCapabilities;
 import org.freedesktop.dbus.exceptions.DBusExecutionException;
 import org.freedesktop.dbus.interfaces.Properties;
 import org.freedesktop.dbus.types.UInt32;
