+++ kura/test/org.eclipse.kura.nm.test/src/test/java/org/eclipse/kura/nm/NMDeviceTypeTest.java
@@ -14,6 +14,7 @@ package org.eclipse.kura.nm;
 
 import static org.junit.Assert.assertEquals;
 
+import org.eclipse.kura.nm.enums.NMDeviceType;
 import org.freedesktop.dbus.types.UInt32;
 import org.junit.Test;
